+++ run_dcr.sh
@@ -15,6 +15,7 @@ if [ -z "$1" ]; then
     echo "=============================================================================="
     echo "all   - Run the complete processing of all new documents."
     echo "db_c  - Create the database."
+    echo "db_u  - Upgrade the database.""
     echo "m_d   - Run the installation of the necessary 3rd party packages for development and run the development ecosystem."
     echo "m_p   - Run the installation of the necessary 3rd party packages for production and compile all packages and modules."
     echo "p_i   - Process the inbox directory."
@@ -71,7 +72,7 @@ case "${DCR_CHOICE_ACTION}" in
         exit 255
     fi
     ;;
-  all|db_c|p_i|p_2_i)
+  all|db_c|db_u|p_i|p_2_i)
     if ! ( pipenv run python src/dcr/dcr.py "${DCR_CHOICE_ACTION}" ); then
         exit 255
     fi