+++ run_test.sh
@@ -15,6 +15,7 @@ if [ -z "$1" ]; then
     echo "=============================================================================="
     echo "all   - Run the complete processing of all new documents."
     echo "db_c  - Create the database."
+    echo "db_u  - Upgrade the database.""
     echo "m_d   - Run the installation of the necessary 3rd party packages for development and run the development ecosystem."
     echo "m_p   - Run the installation of the necessary 3rd party packages for production and compile all packages and modules."
     echo "p_i   - Process the inbox directory."
@@ -67,6 +68,9 @@ case "${DCR_CHOICE_ACTION}" in
     rm -f data/dcr.db
     pipenv run python src/dcr/dcr.py "${DCR_CHOICE_ACTION}"
     ;;
+  db_u|p_2_i)
+    pipenv run python src/dcr/dcr.py "${DCR_CHOICE_ACTION}"
+    ;;
   all|p_i)
     rm -rf data/inbox
     mkdir data/inbox
@@ -74,7 +78,7 @@ case "${DCR_CHOICE_ACTION}" in
     pipenv run python src/dcr/dcr.py "${DCR_CHOICE_ACTION}"
     ;;
   *)
-    echo "Usage: ./run_test.sh all | db_c | m_d | m_p | p_i | p_2_i"
+    echo "Usage: ./run_test.sh all | db_c | db_u | m_d | m_p | p_i | p_2_i"
     ;;
 esac